other rest-controller/services, adding of using transactions in services, fix Integer to Long

+++ src/main/java/com/inventory/api/model/EquipmentUnit.java
@@ -17,7 +17,7 @@ public class EquipmentUnit {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "equipment_unit_id")
-    private Integer equipmentUnitId;
+    private Long equipmentUnitId;
 
     @ManyToOne
     @JoinColumn(name = "equipment_id")
